fix: avoid side-effects of helper imports
