+++ packages/nuxt/src/core/utils/names.ts
@@ -1,0 +1,14 @@
+import { basename, extname } from 'pathe'
+import { kebabCase, pascalCase } from 'scule'
+
+export function getNameFromPath (path: string) {
+  return kebabCase(basename(path).replace(extname(path), '')).replace(/["']/g, '')
+}
+
+export function hasSuffix (path: string, suffix: string) {
+  return basename(path).replace(extname(path), '').endsWith(suffix)
+}
+
+export function getImportName (name: string) {
+  return pascalCase(name).replace(/[^\w]/g, r => '_' + r.charCodeAt(0))
+}
